Add explicit match arms for advanced eval_univariate
Added explicit match arms for eval univariate

Changed extended polynomial tests to intermediate polynomial tests
